% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_colour_palette}
\alias{default_colour_palette}
\title{Hex codes of all colours in tailwind css default colour palette}
\format{
A named list of length 82.
}
\source{
\url{https://tailwindcss.com/docs/customizing-colors/#default-color-palette}
}
\usage{
default_colour_palette
}
\description{
A dataset containing 82 colours.
}
\keyword{datasets}
