% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_tailwind2.R
\name{scale_colour_tailwind2}
\alias{scale_colour_tailwind2}
\title{Use tailwind palettes to colour ggplot2 plots}
\usage{
scale_colour_tailwind2(palette, discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette function, use \code{tailwind_pal_scale}}

\item{discrete}{Boolean, is the colour variable categorical/discrete}

\item{reverse}{Boolean, whether you want to flip palette order}

\item{...}{Additional options to pass to \code{discrete_scale} or \code{scale_color_gradientn}}
}
\description{
Use tailwind palettes to colour ggplot2 plots
}
