% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_tailwind.R
\name{scale_fill_tailwind}
\alias{scale_fill_tailwind}
\title{Use tailwind palettes to colour ggplot2 plots}
\usage{
scale_fill_tailwind(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name from \code{tailwind_palettes}}

\item{discrete}{Boolean, is the colour variable categorical/discrete}

\item{reverse}{Boolean, whether you want to flip palette order}

\item{...}{Additional options to pass to \code{discrete_scale} or \code{scale_fill_gradientn}}
}
\description{
Use tailwind palettes to colour ggplot2 plots
}
