% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailwind_cols.R
\name{tailwind_cols}
\alias{tailwind_cols}
\title{Extract hex codes from default colour pallete}
\usage{
tailwind_cols(...)
}
\arguments{
\item{...}{Character names of tailwind colours}
}
\description{
Extract hex codes from default colour pallete
}
\examples{
tailwind_cols()
tailwind_cols('pink-500')
tailwind_cols('red-700', 'blue-500')

}
