% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailwind_cols_regex.R
\name{tailwind_cols_regex}
\alias{tailwind_cols_regex}
\title{Extract hex codes using regex pattern}
\usage{
tailwind_cols_regex(pattern)
}
\arguments{
\item{pattern}{Regex pattern to search palette names}
}
\description{
Extract hex codes using regex pattern
}
\examples{
tailwind_cols_regex('yellow-.*')
tailwind_cols_regex('.*-400')

}
