% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailwind_pal.R
\name{tailwind_pal}
\alias{tailwind_pal}
\title{Create palette function using tailwind_palettes}
\usage{
tailwind_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name from \code{tailwind_palettes}}

\item{reverse}{Boolean, whether you want to flip palette order}

\item{...}{Additional options to pass to \code{colorRampPalette}}
}
\description{
This is a function to generate a palette function from \code{tailwind_palettes} for us with \code{discrete_scale} and \code{ggplot} functions.
}
