% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailwind_pal_scale.R
\name{tailwind_pal_scale}
\alias{tailwind_pal_scale}
\title{Create a scale palette using tailwind colours}
\usage{
tailwind_pal_scale(start = "green-500", end = "red-500",
  middle = NULL, ...)
}
\arguments{
\item{start}{Character name from \code{tailwind_palettes} for the low end of the resulting scale}

\item{end}{Character name from \code{tailwind_palettes} for the high end of the resulting scale}

\item{middle}{Optional character name from \code{tailwind_palettes} for the scale to go through}

\item{...}{Additional options to pass to \code{colorRampPalette}}
}
\description{
Generate a palette function object with a start hex, end hex and optional middle colour that the scale will go through.
}
