% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\docType{data}
\name{tailwind_palettes}
\alias{tailwind_palettes}
\title{A collection of basic colour palettes}
\format{A named list of length 20}
\usage{
tailwind_palettes
}
\description{
Combinations of colours found in \code{default_color_palette}. Split by base colour and shade.
}
\seealso{
\code{default_colour_palette}
}
\keyword{datasets}
